callout creation details & template steps
