+++ src/domain/callout/creation-dialog/steps/CalloutInfoStep.tsx
@@ -3,20 +3,22 @@ import { useTranslation } from 'react-i18next';
 import { Box } from '@mui/material';
 import CalloutForm, { CalloutFormOutput } from '../../../../components/composite/aspect/AspectCreationDialog/form/CalloutForm';
 import { CalloutCreationType } from '../CalloutCreationDialog';
-import { StepComponentProps } from '../../../shared/components/Stepper/step/Step';
+import { StepComponentProps } from '../../../shared/components/Steps/step/Step';
 import { StepLayoutImpl } from '../step-layout/StepLayout';
+
 export interface CalloutInfoStepProps {
   callout: CalloutCreationType;
   onChange?: (callout: CalloutFormOutput) => void;
   onStatusChanged?: (isValid: boolean) => void;
+  onClose?: () => void;
 }
 
-const CalloutInfoStep: FC<StepComponentProps & CalloutInfoStepProps> = ({ callout, onChange, onStatusChanged, activeStep, isValid, steps, next, prev }) => {
+const CalloutInfoStep: FC<StepComponentProps & CalloutInfoStepProps> = ({ callout, onChange, onClose, onStatusChanged, activeStep, isValid, steps, next, prev }) => {
   const { t } = useTranslation();
   return (
     <StepLayoutImpl
       dialogTitle={t('components.callout-creation.title')}
-      onClose={() => {}}
+      onClose={onClose}
       next={next} prev={prev}
       activeStep={activeStep}
       steps={steps}
