+++ src/domain/callout/creation-dialog/steps/CalloutTemplateStep/CalloutTemplateStepProps.ts
@@ -1,0 +1,7 @@
+import { CalloutCreationType } from '../../CalloutCreationDialog';
+
+export interface CalloutTemplateStepProps {
+  callout: CalloutCreationType;
+  onChange?: (templateId: string) => void;
+  onClose?: () => void;
+}
