+++ src/domain/shared/components/Steps/Steps.tsx
@@ -52,6 +52,7 @@ const Steps = <PassedProps extends {}>({ initialActiveStep, children: childrenOr
 
   // Taking whatever props we make use of here from a <Step>, passing the rest to the step Component
   const { component: Component, title, ...passedProps } = currentStep.props;
+  console.log(passedProps);
 
   return <Component activeStep={activeStep} steps={definitions} prev={prev} next={next} {...passedProps as PassedProps} />;
 };
