+++ src/pages/Contribute/ContributePage.tsx
@@ -1,4 +1,4 @@
-import React, { FC, useMemo } from 'react';
+import React, { FC, useCallback, useMemo, useState } from 'react';
 import { useUrlParams } from '../../hooks';
 import ContributeTabContainer from '../../containers/ContributeTabContainer/ContributeTabContainer';
 import ContributeView from '../../views/ContributeView/ContributeView';
@@ -10,12 +10,17 @@ import OpportunityPageLayout from '../../domain/opportunity/layout/OpportunityPa
 import CanvasesView from '../../domain/canvas/EntityCanvasPage/CanvasesView';
 import { useResolvedPath } from 'react-router-dom';
 import { SectionSpacer } from '../../domain/shared/components/Section/Section';
+import { Button } from '@mui/material';
+import { useTranslation } from 'react-i18next';
+import Box from '@mui/material/Box';
+import CalloutCreationDialog from '../../domain/callout/creation-dialog/CalloutCreationDialog';
 
 interface ContributePageProps {
   entityTypeName: EntityTypeName;
 }
 
 const ContributePage: FC<ContributePageProps> = ({ entityTypeName }) => {
+  const { t }  = useTranslation();
   const { hubNameId, challengeNameId, opportunityNameId, canvasId } = useUrlParams();
 
   const currentPath = useResolvedPath(canvasId ? '..' : '.');
@@ -32,21 +37,41 @@ const ContributePage: FC<ContributePageProps> = ({ entityTypeName }) => {
     throw new TypeError(`Unknown entity ${entityTypeName}`);
   }, [entityTypeName]);
 
+  const [isCalloutCreationDialogOpen, setIsCalloutCreationDialogOpen] = useState(true);
+  const handleCreateCalloutOpened = useCallback(() => setIsCalloutCreationDialogOpen(true), []);
+  const handleCreateCalloutClosed = useCallback(() => setIsCalloutCreationDialogOpen(false), []);
+  const handleCreateCalloutCreated = useCallback(async () => { return undefined; }, []);
+
   if (!hubNameId) {
     return <></>;
   }
 
   return (
-    <PageLayout currentSection={EntityPageSection.Explore}>
-      <CanvasesView canvasId={canvasId} parentUrl={currentPath.pathname} entityTypeName={entityTypeName} />
-      <SectionSpacer />
-      <ContributeTabContainer
-        hubNameId={hubNameId}
-        challengeNameId={challengeNameId}
-        opportunityNameId={opportunityNameId}
-        component={ContributeView}
+    <>
+      <PageLayout currentSection={EntityPageSection.Explore}>
+        <Box display="flex" justifyContent="end">
+          <Button
+            variant="contained"
+            onClick={handleCreateCalloutOpened}
+          >
+            {t('pages.explore.create-callout')}
+          </Button>
+        </Box>
+        <CanvasesView canvasId={canvasId} parentUrl={currentPath.pathname} entityTypeName={entityTypeName} />
+        <SectionSpacer />
+        <ContributeTabContainer
+          hubNameId={hubNameId}
+          challengeNameId={challengeNameId}
+          opportunityNameId={opportunityNameId}
+          component={ContributeView}
+        />
+      </PageLayout>
+      <CalloutCreationDialog
+        open={isCalloutCreationDialogOpen}
+        onClose={handleCreateCalloutClosed}
+        onCreate={handleCreateCalloutCreated}
       />
-    </PageLayout>
+    </>
   );
 };
